move test, clean it up
