--- test/e2e_node/terminating_pod_test.go
@@ -1,137 +1,0 @@
-/*
-Copyright 2024 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package e2enode
-
-import (
-	"context"
-	"fmt"
-	"os"
-	"path/filepath"
-	"time"
-
-	v1 "k8s.io/api/core/v1"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/uuid"
-	clientset "k8s.io/client-go/kubernetes"
-	"k8s.io/kubernetes/test/e2e/framework"
-	imageutils "k8s.io/kubernetes/test/utils/image"
-	admissionapi "k8s.io/pod-security-admission/api"
-
-	"github.com/onsi/ginkgo/v2"
-	"github.com/onsi/gomega"
-)
-
-var _ = SIGDescribe("TerminatingPod", func() {
-	f := framework.NewDefaultFramework("terminating-pod")
-	f.NamespacePodSecurityEnforceLevel = admissionapi.LevelBaseline
-	ginkgo.Context("when kubelet restarts and there is a terminating pod", func() {
-		var ns, podPath, podName string
-		ginkgo.It("should finish terminating", func(ctx context.Context) {
-			ns = f.Namespace.Name
-			podName = "terminating-pod-" + string(uuid.NewUUID())
-
-			ginkgo.By("create the pod with a deletion timestamp")
-			err := createTerminatingPod(podPath, podName, ns, imageutils.GetE2EImage(imageutils.Nginx))
-			framework.ExpectNoError(err)
-
-			ginkgo.By("wait for the pod to be terminating")
-			gomega.Eventually(ctx, func(ctx context.Context) error {
-				return verifyPodTerminating(ctx, f.ClientSet, podName, ns)
-			}, 2*time.Minute, time.Second*1).Should(gomega.BeNil())
-
-			ginkgo.By("wait for the pod to completely terminate")
-			gomega.Eventually(ctx, func(ctx context.Context) error {
-				return verifyPodDisappeared(ctx, f.ClientSet, podName, ns)
-			}, 2*time.Minute, time.Second*1).Should(gomega.BeNil())
-		})
-	})
-})
-
-func makePodPath(dir, name, namespace string) string {
-	return filepath.Join(dir, namespace+"-"+name+".yaml")
-}
-
-func createTerminatingPod(dir, name, namespace, image string) error {
-	template := `
-apiVersion: v1
-kind: Pod
-metadata:
-  name: %s
-  namespace: %s
-  deletionTimestamp: "2024-05-16T01:29:53Z"
-spec:
-  containers:
-  - name: test
-    image: %s
-    lifecycle:
-      preStop:
-        exec:
-          command:
-          - "sleep"
-          - "10"
-    volumeMounts:
-    - mountPath: /kube-certs
-      name: kube-ssl
-      readOnly: true
-  volumes:
-  - name: kube-ssl
-    secret:
-      defaultMode: 420
-      secretName: kube-public-cert
-`
-	file := makePodPath(dir, name, namespace)
-	podYaml := fmt.Sprintf(template, name, namespace, image)
-
-	f, err := os.OpenFile(file, os.O_RDWR|os.O_TRUNC|os.O_CREATE, 0666)
-	if err != nil {
-		return err
-	}
-	defer func(f *os.File) {
-		_ = f.Close()
-	}(f)
-
-	_, err = f.WriteString(podYaml)
-	return err
-}
-
-func verifyPodTerminating(ctx context.Context, cl clientset.Interface, name, namespace string) error {
-	pod, err := cl.CoreV1().Pods(namespace).Get(ctx, name, metav1.GetOptions{})
-	if err != nil {
-		return fmt.Errorf("expected the pod %q to appear: %w", name, err)
-	}
-	if pod.Status.Phase != v1.PodRunning {
-		return fmt.Errorf("expected the pod %q to be running, got %q", name, pod.Status.Phase)
-	}
-	for i := range pod.Status.ContainerStatuses {
-		if pod.Status.ContainerStatuses[i].State.Running == nil {
-			return fmt.Errorf("expected the pod %q with container %q to be running (got containers=%v)", name, pod.Status.ContainerStatuses[i].Name, pod.Status.ContainerStatuses[i].State)
-		}
-	}
-	return nil
-}
-
-func verifyPodDisappeared(ctx context.Context, cl clientset.Interface, name, namespace string) error {
-	_, err := cl.CoreV1().Pods(namespace).Get(ctx, name, metav1.GetOptions{})
-	if apierrors.IsNotFound(err) {
-		return nil
-	}
-	if err == nil {
-		return fmt.Errorf("pod %v/%v still exists", namespace, name)
-	}
-	return fmt.Errorf("expected pod %v/%v to not exist but got error: %w", namespace, name, err)
-}
